refactor(app): move AppSettings pages
