+++ app/src/App/DesktopApp.tsx
@@ -23,7 +23,7 @@ import { ProtocolRunDetails } from '../pages/Desktop/Devices/ProtocolRunDetails'
 import { RobotSettings } from '../pages/Desktop/Devices/RobotSettings'
 import { ProtocolsLanding } from '../pages/Desktop/Protocols/ProtocolsLanding'
 import { ProtocolDetails } from '../pages/Desktop/Protocols/ProtocolDetails'
-import { AppSettings } from '../pages/AppSettings'
+import { AppSettings } from '../pages/Desktop/AppSettings'
 import { Labware } from '../pages/Desktop/Labware'
 import { useSoftwareUpdatePoll } from './hooks'
 import { Navbar } from './Navbar'
