+++ app/src/pages/Desktop/AppSettings/index.tsx
@@ -15,15 +15,15 @@ import {
   TYPOGRAPHY,
 } from '@opentrons/components'
 
-import * as Config from '../../redux/config'
+import * as Config from '../../../redux/config'
 import { GeneralSettings } from './GeneralSettings'
 import { PrivacySettings } from './PrivacySettings'
 import { AdvancedSettings } from './AdvancedSettings'
-import { FeatureFlags } from '../../organisms/AppSettings/FeatureFlags'
-import { NavTab } from '../../molecules/NavTab'
-import { Line } from '../../atoms/structure'
+import { FeatureFlags } from '../../../organisms/AppSettings/FeatureFlags'
+import { NavTab } from '../../../molecules/NavTab'
+import { Line } from '../../../atoms/structure'
 
-import type { DesktopRouteParams, AppSettingsTab } from '../../App/types'
+import type { DesktopRouteParams, AppSettingsTab } from '../../../App/types'
 
 export function AppSettings(): JSX.Element {
   const { t } = useTranslation('app_settings')
